pyproject.toml
[project]
name = "service-ask-marc"
version = "0.1.0"
description = "A bridge for OpenRemote to communicate with AI models like ChatGPT or Claude. And provide configuration to connect them to tools using MCP."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.0",
    "fastmcp==2.13.0.2",
    "langchain>=1.0.5",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=1.0.2",
    "openremote-client>=1.0.2",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]
